[gdb/symtab] Work around a gcc PR fixed in gcc 9
When running test-case gdb.base/bg-execution-repeat.exp with target board
dwarf5-fission-debug-types, I get:
...
(gdb) break -qualified main^M
Offset from DW_FORM_GNU_str_index or DW_FORM_strx pointing outside of \
  .debug_str.dwo section in CU at offset 0x0 [in module bg-execution-repeat]^M
Offset from DW_FORM_GNU_str_index or DW_FORM_strx pointing outside of \
  .debug_str.dwo section in CU at offset 0x0 [in module bg-execution-repeat]^M
(gdb) FAIL: $exp: c&: gdb_breakpoint: set breakpoint at main
...

This happens with gcc 7 and 8, but not with gcc 9 and later.

We're using dwarf5, so the .debug_str_offsets section is supposed to have a
header, but there's none:
...
        .section        .debug_str_offsets.dwo,"e",@progbits
        .long   0       # indexed string 0x0: do_wait
        .long   0x8     # indexed string 0x1: unsigned int
        .long   0x15    # indexed string 0x2: wait
...

In other words, we have a v4-style .debug_str_offsets section.

Workaround this by detecting that it's not a valid v5 header by comparing the
initial length to the the section size, and handling it as a v4 section.

IWBN to enable this workaround only for gcc version < 9, but unfortunately the
workaround is required to correctly read the producer string, so that's not
possible.

Tested on x86_64-linux.

PR symtab/32861
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=32861