mingw-hdep: Add "maint set testuite-mode on/off" command
This patch is to add

  maint set testsuite mode on/off

command.

Finally, thanks to Yao's patch about handling of flushing of stderr for mingw,
it reduces to switching the stdout and stderr between binary and text mode.

Setting the file handles to binary mode avoid all the trouble about double ^M
that confuses DejaGnu so often when trying to run the testsuite with a mingw
compiled GDB.