[gdb/testsuite] Set TERM to dumb by default
With MSYS2 and default TERM=xterm-256color (as well as with xterm and ansi), I
get:
...
builtin_spawn gdb -q ...
^[[6n(gdb) ERROR: GDB never initialized.
...

This is not specific to gdb, other tools produce the same CSI sequence, and
consequently we run into trouble in other places (like get_compiler_info).

Fix this by default-setting TERM to dumb.

We do this for all platforms, to avoid test-cases passing on one platform but
failing on another.

For test-case that set TERM to something other than dumb, handle the CSI
sequence in default_gdb_start.

PR testsuite/33072
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=33072