[gdb/testsuite] Support clean_restart $binfile
When running test-case gdb.base/access-mem-running.exp, we do:
...
      clean_restart ${binfile}
...
with binfile set to absolute path "$dir/access-mem-running".

The proc documentation states that the executable argument is a basename
though, so it does:
...
	set binfile [standard_output_file ${executable}]
...
and consequently in standard_output_file we do the equivalent of
"file join $dir $dir/access-mem-running".

This doesn't return $dir/$dir/access-mem-running though, but simply
$dir/access-mem-running.  The tcl command file join
discards all arguments preceding an absolute path.

However, that doesn't work on msys2, where dir is of the form "C:/...", and
we do get $dir/$dir/access-mem-running.

Confusingly, doing the same in a standalone tcl script works fine, so this may be
a configuration or porting problem.

We could work around this in standard_output_file by detecting the
<drive-letter>:/ prefix in basename, as is done here [1].

Instead, handle this in a platform-independent way, by checking for
"regexp ^[string_to_regexp $dir] $basename".

[1] https://github.com/ssbssa/gdb/commit/93433da5a556e3e7956692a41e8ad22db6ab67f7