[gdb/testsuite] Support clean_restart $binfile 65;7402;1c When running test-case gdb.base/access-mem-running.exp, we do: ...       clean_restart ${binfile} ... with binfile set to absolute path "$dir/access-mem-running".
The proc documentation states that the executable argument is a basename
though, so it does:
...
	set binfile [standard_output_file ${executable}]
...
and consequently in standard_output_file we do the equivalent of
"file join $dir $dir/access-mem-running".

This doesn't return $dir/$dir/access-mem-running though, but simply
$dir/access-mem-running.  The tcl command "file join"
discards all arguments preceding an absolute path.

However, that doesn't work on msys2, where dir is of the form "C:/...", and
we do get $dir/$dir/access-mem-running.

[ Confusingly, doing the same in a standalone tcl script works fine, so this
may be a configuration or porting problem. ]

Fix this by no longer relying on "file join" to solve this.

Instead, check that the basename argument in standard_output_file is not an
absolute filename.

Allow though the exception that the basename argument is an absolute file name
in the standard output dir, since this is both common and harmless.

Checking absolute file name using "file pathname" doesn't work on mingw, so
introduce a new proc isabsolute.

That addresses the "clean_restart $binfile" case, but clean_restart is also
used with executables from the standard temp dir, which triggers the newly
introduced check.

Fix this by allowing absolute filenames in clean_restart, and only adding the
standard output dir prefix for relative filenames.

The stricter checking also triggers in get_build_id, so fix that as well.

Tested on x86_64-linux.