[gdb/symtab] Bail out of create_addrmap_from_gdb_index on error
As mentioned in the previous commit, when running test-case
gdb.tui/tui-missing-src.exp with target board gold-gdb-index (and fission and
fission-dwp) on aarch64-linux, I run into:
...
FAIL: gdb.tui/tui-missing-src.exp: checking if inside f2 ()
...

The address table of the .gdb_index section contains an entry for function f2
with an invalid address range.

While the previous commit:
- allows us to detect the problem using complaints, and
- skips the invalid address range
the incorrect index is still used, and the test-case still fails because it
needs the correct address range for f2.

Fix this by bailing out of create_addrmap_from_gdb_index on finding errors in
the address table.

This fixes the test-case for the 3 target boards for which this was failing.

Overall, this reduces the amount of failing test-cases with target board
gold-gdb-index on aarch64-linux from 27 to 12.

There are also two regressions:
- gdb.base/gdb-index-err.exp
- gdb.base/index-cache.exp

In both cases, the incorrect index is now rejected and this makes the
test-case fail.

Tested on aarch64-linux.