[gdb/symtab] Turn complaints in create_addrmap_from_gdb_index into warnings
Rather than issuing a complaint, which is off by default, warn when returning
false in create_addrmap_from_gdb_index, informing the user that the .gdb_index
was ignored, and why.

This increases the amount of failing test-cases with target board
gold-gdb-index on aarch64-linux from 12 to 27.

Tested on aarch64-linux.