[gdb/symtab] Refactor warnings in create_addrmap_from_gdb_index
After introducing the invalid_range_warning lambda function in
create_addrmap_from_gdb_index (suggested by Simon), I realized that by adding
a reason argument to it:
- I could use it in more places, and
- add more information in the cases currently using it.

Do so, and while we're at it, add a similar lambda function
invalid_index_warning.  It's used only once, but it makes the structure of the
function more regular.

Tested on x86_64-linux.