[gdb/testsuite] Remove wait in gdb.tui/pr30056.exp
In test-case gdb.tui/pr30056.exp, we have:
...
    # Send arrow-right.
    send_gdb "\033\[C"

    # We need to the arrow-right to be processed by readline, before we
    # send the following ^C.  Waiting 1 ms seems to do that.
    after 1
...

Waiting for an effect by sleeping is fragile.

The reason we're doing so is that we're waiting for an effect that is not
visible: the arrow-right is meant to horizontally scroll the source window,
but since there's no source, nothing happens.

Fix this by using an executable in the test-case, with a source file
containing a comment long enough to make horizontal scrolling possible.

Tested on x86_64-linux.