[gdb/build] Rename .c files to .cc
Rename the .c files in gdb to .cc.  We do this in a few steps.

First, do the rename in a script.
...
 #!/bin/bash

 files=($(find gdb -type f -name "*.c" | egrep -v /testsuite/))
 for f in "${files[@]}"; do
     git mv "$f" "${f}c"
 done
...

Then update the includes to use the new names.  Likewise for *.y:
...
$ find gdb -type f -name "*.cc" \
  | grep -v /testsuite/ \
  | xargs sed -i 's/#include "\(.*\.c\)"/#include "\1c"/'
...

Then update gdb/Makefile.in, gdb/features/Makefile and gdbserver/Makefile.in
to use the new names.

Tested on x86_64-linux.