[gdb/contrib] Reduce check-gnu-style false positives
The recently introduced pre-commit hook check-gnu-style uses script
contrib/check_GNU_style.py.  The script takes as input a patch file.

Patch files may not provide sufficient context to determine whether the
text being added is code, a string literal, a char literal or a comment.

Consequently, a check that should only trigger on code, also triggers on
comments and strings:
...
$ cat tmp.patch
  ...
+extern int foo();
+char str[] = "foo()";
+/* foo() */
$ ./contrib/check_GNU_style.py tmp.patch
=== ERROR type #1: there should be exactly one space between function name \
  and parenthesis (3 error(s)) ===
test.c:1:14:extern int foo();
test.c:2:17:char str[] = "foo()";
test.c:3:6:/* foo() */
$
...

However, in the case of the pre-commit check check-gnu-style, we have access
to both:
- the patch, and
- the files being patched.

Add a script gdb/contrib/filter-c.py that can be used to filter out comments
and strings literals (as well as char literals while we're at it) like this:
...
$ gdb/contrib/filter-c.py test.c
extern int foo();
char str[] = "sssss";
/* CCCCC */
...
and use it to recreate the patch before running ./contrib/check_GNU_style.py on
it.

Consequently, we only see the error for the code case:
...
$ git commit -m try
  ...
check-gnu-style..........................................................Passed
- hook id: check-gnu-style
- duration: 0.06s

=== ERROR type #1: there should be exactly one space between function name \
  and parenthesis (1 error(s)) ===
gdb/test.c.post.filtered:1:14:extern int foo();
  ...
$
...

The filter script gdb/contrib/filter-c.py tries to be a bit smart about
how it filters out various parts.

In the case of strings, we don't want any check to trigger, so we replace
every single char.

But in the case of comments, we still want to check whitespace and end of line
formatting, so whitespace and dots are left intact.

An alternative could be to generate two versions of each file:
- one containing only comments, and
- one containing only code
and to do only the appropriate checks on each file.

Or, take the integration one step further and do the checks immediately
while parsing, eliminating the need to write out files.