[contrib] Fix SentenceSeparatorCheck in check_GNU_style_lib.py
The SentenceSeparatorCheck in check_GNU_style_lib.py warns about too few (1)
or too many (3+) spaces in between sentences:
...
/* First sentence. Second sentence.  */
/* First sentence.   Second sentence.  */
...

But it also warns about abbreviations, like 'i.e.' or 'ie.'.

Fix this, and add a corresponding regression test.

We recognize any abbreviation of the form 'i.e.'.  For abbreviations like
'ie.', we use a list, because they're indistinguishable from words at the end
of a sentence.

[1] https://www.gnu.org/prep/standards/html_node/Comments.html
[2] https://www.gnu.org/software/emacs/manual/html_node/emacs/Sentences.html