[gdb/testsuite] Fix gdb.base/inline-frame-cycle-unwind.exp for s390x (alternative)
With test-case gdb.base/inline-frame-cycle-unwind.exp on s390x-linux, I run
into:
...
 (gdb) bt^M
 #0  inline_func () at inline-frame-cycle-unwind.c:49^M
 #1  normal_func () at inline-frame-cycle-unwind.c:32^M
 #2  0x000000000100065c in inline_func () at inline-frame-cycle-unwind.c:45^M
 #3  normal_func () at inline-frame-cycle-unwind.c:32^M
 Backtrace stopped: previous frame identical to this frame (corrupt stack?)^M
 (gdb) FAIL: $exp: bt: cycle at level 5: backtrace when the unwind is broken \
   at frame 5
...

In contrast, on x86_64-linux, I get:
...
 (gdb) bt^M
 #0  inline_func () at inline-frame-cycle-unwind.c:49^M
 #1  normal_func () at inline-frame-cycle-unwind.c:32^M
 #2  0x0000000000401157 in inline_func () at inline-frame-cycle-unwind.c:45^M
 #3  normal_func () at inline-frame-cycle-unwind.c:32^M
 #4  0x0000000000401157 in inline_func () at inline-frame-cycle-unwind.c:45^M
 #5  normal_func () at inline-frame-cycle-unwind.c:32^M
 Backtrace stopped: previous frame identical to this frame (corrupt stack?)^M
 (gdb) PASS: $exp: bt: cycle at level 5: backtrace when the unwind is broken \
   at frame 5
...

AFAIU, the mechanism of the test is as follows: the custom unwinder produces the
frame-id for frame #5 at frame #4.  Consequently, when arriving at frame #5, a
cycle is detected.

This relies on the test-case to calculate the offending frame-id, and the
problem on s390x is that that calculation is incorrect.

Fix this by using "maint print frame-id" to get all frame-ids.

Tested on x86_64-linux and s390x-linux.